@@ -1,39 +1,39 @@
 {
-  "_from": "ms@2.0.0",
-  "_id": "ms@2.0.0",
+  "_from": "ms@^2.1.1",
+  "_id": "ms@2.1.2",
   "_inBundle": false,
-  "_integrity": "sha1-VgiurfwAvmwpAd9fmGF4jeDVl8g=",
+  "_integrity": "sha512-sGkPx+VjMtmA6MX27oA4FBFELFCZZ4S4XqeGOXCv68tT+jb3vk/RyaKWP0PTKyWtmLSM0b+adUTEvbs1PEaH2w==",
   "_location": "/ms",
   "_phantomChildren": {},
   "_requested": {
-    "type": "version",
+    "type": "range",
     "registry": true,
-    "raw": "ms@2.0.0",
+    "raw": "ms@^2.1.1",
     "name": "ms",
     "escapedName": "ms",
-    "rawSpec": "2.0.0",
+    "rawSpec": "^2.1.1",
     "saveSpec": null,
-    "fetchSpec": "2.0.0"
+    "fetchSpec": "^2.1.1"
   },
   "_requiredBy": [
     "/debug"
   ],
-  "_resolved": "https://registry.npmjs.org/ms/-/ms-2.0.0.tgz",
-  "_shasum": "5608aeadfc00be6c2901df5f9861788de0d597c8",
-  "_spec": "ms@2.0.0",
-  "_where": "C:\\NodeJS\\MutantTest-Node.js\\node_modules\\debug",
+  "_resolved": "https://registry.npmjs.org/ms/-/ms-2.1.2.tgz",
+  "_shasum": "d09d1f357b443f493382a8eb3ccd183872ae6009",
+  "_spec": "ms@^2.1.1",
+  "_where": "D:\\NodeJS\\node_modules\\debug",
   "bugs": {
     "url": "https://github.com/zeit/ms/issues"
   },
   "bundleDependencies": false,
   "deprecated": false,
-  "description": "Tiny milisecond conversion utility",
+  "description": "Tiny millisecond conversion utility",
   "devDependencies": {
-    "eslint": "3.19.0",
+    "eslint": "4.12.1",
     "expect.js": "0.3.1",
-    "husky": "0.13.3",
-    "lint-staged": "3.4.1",
-    "mocha": "3.4.1"
+    "husky": "0.14.3",
+    "lint-staged": "5.0.0",
+    "mocha": "4.0.1"
   },
   "eslintConfig": {
     "extends": "eslint:recommended",
@@ -65,5 +65,5 @@
     "precommit": "lint-staged",
     "test": "mocha tests.js"
   },
-  "version": "2.0.0"
+  "version": "2.1.2"
 }
